backend/requirements.txt
fastapi
uvicorn
requests
faiss-cpu
python-dotenv
sentence-transformers
langchain
